Get rid of code duplucation after adding CHECK GRANT.
