+++ src/Interpreters/Access/InterpreterCheckGrantQuery.h
@@ -1,16 +1,11 @@
 #pragma once
 
-#include <Core/UUID.h>
 #include <Interpreters/IInterpreter.h>
 #include <Parsers/IAST_fwd.h>
 
 
 namespace DB
 {
-
-class ASTCheckGrantQuery;
-struct User;
-struct Role;
 
 class InterpreterCheckGrantQuery : public IInterpreter, WithMutableContext
 {
